requirements.txt
fastapi~=0.115.8
uvicorn[standard]==0.23.2
SQLAlchemy~=2.0.38
python-dotenv==1.0.0
pydantic-settings==2.5.2
email-validator==2.3.0
pandas==2.3.3
psycopg[binary]>=3.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pwdlib[argon2]>=0.2.0
PyJWT>=2.8.0
boto3
pydantic>=2
numpy==2.2.4
openai>=2.7.1
opencv-python-headless==4.12.0.88
pillow>=12.0.0
pydantic>=2.12.4
python-multipart>=0.0.20
python-dotenv
pytesseract
pillow

requests
google-auth>=2.27.0
openrouter==0.0.19
httpx>=0.24.0
gunicorn>=21.2.0
